requirements.txt
flask
grpcio
grpcio-tools
kafka-python